Split Matrixdock fees adapters
